+++ config/config.go
@@ -47,7 +47,7 @@ type (
 
 		Agent struct {
 			Token       string
-			Image       string `default:"drone/agent:1"`
+			Image       string `default:"drone/drone-runner-docker:1"`
 			Concurrency int    `default:"2"`
 			OS          string `default:"linux"`
 			Arch        string `default:"amd64"`